[SYNTH-12989] Unexport `getTestAndOverrideConfig()`
