+++ src/commands/synthetics/__tests__/utils/public.test.ts
@@ -320,127 +320,6 @@ describe('utils', () => {
     })
   })
 
-  describe('getTestAndOverrideConfig', () => {
-    test('Forbidden error when getting a test', async () => {
-      const axiosMock = jest.spyOn(axios, 'create')
-      axiosMock.mockImplementation((() => (e: any) => {
-        throw getAxiosError(403, {message: 'Forbidden'})
-      }) as any)
-
-      const triggerConfig = {suite: 'Suite 1', config: {}, id: '123-456-789'}
-
-      await expect(() =>
-        utils.getTestAndOverrideConfig(api, triggerConfig, mockReporter, getSummary())
-      ).rejects.toThrow('Failed to get test: could not query https://app.datadoghq.com/example\nForbidden\n')
-    })
-
-    test('Passes when public ID is valid', async () => {
-      const axiosMock = jest.spyOn(axios, 'create')
-      axiosMock.mockImplementation((() => (e: any) => {
-        return {data: {subtype: 'http', public_id: '123-456-789'}}
-      }) as any)
-
-      const triggerConfig = {suite: 'Suite 1', config: {}, id: '123-456-789'}
-      expect(await utils.getTestAndOverrideConfig(api, triggerConfig, mockReporter, getSummary())).toEqual(
-        expect.objectContaining({test: expect.objectContaining({public_id: '123-456-789', subtype: 'http'})})
-      )
-    })
-
-    test('Fails when public ID is NOT valid', async () => {
-      const expectedError = new CiError('INVALID_CONFIG', `No valid public ID found in: \`a123-456-789\``)
-
-      const triggerConfig = {suite: 'Suite 1', config: {}, id: 'a123-456-789'}
-      await expect(utils.getTestAndOverrideConfig(api, triggerConfig, mockReporter, getSummary())).rejects.toThrow(
-        expectedError
-      )
-    })
-
-    test('Passes when the tunnel is enabled for HTTP test', async () => {
-      const axiosMock = jest.spyOn(axios, 'create')
-      axiosMock.mockImplementation((() => (e: any) => {
-        return {data: {subtype: 'http', public_id: '123-456-789'}}
-      }) as any)
-
-      const triggerConfig = {suite: 'Suite 1', config: {}, id: '123-456-789'}
-      expect(await utils.getTestAndOverrideConfig(api, triggerConfig, mockReporter, getSummary(), true)).toEqual(
-        expect.objectContaining({test: expect.objectContaining({public_id: '123-456-789', subtype: 'http'})})
-      )
-    })
-
-    test('Passes when the tunnel is enabled for Browser test', async () => {
-      const axiosMock = jest.spyOn(axios, 'create')
-      axiosMock.mockImplementation((() => (e: any) => {
-        return {data: {type: 'browser', public_id: '123-456-789'}}
-      }) as any)
-
-      const triggerConfig = {suite: 'Suite 1', config: {}, id: '123-456-789'}
-      expect(await utils.getTestAndOverrideConfig(api, triggerConfig, mockReporter, getSummary(), true)).toEqual(
-        expect.objectContaining({test: expect.objectContaining({public_id: '123-456-789', type: 'browser'})})
-      )
-    })
-
-    test('Passes when the tunnel is enabled for Multi step test with HTTP steps only', async () => {
-      const axiosMock = jest.spyOn(axios, 'create')
-      axiosMock.mockImplementation((() => (e: any) => {
-        return {
-          data: {
-            type: 'api',
-            subtype: 'multi',
-            config: {steps: [{subtype: 'http'}, {subtype: 'http'}]},
-            public_id: '123-456-789',
-          },
-        }
-      }) as any)
-
-      const triggerConfig = {suite: 'Suite 1', config: {}, id: '123-456-789'}
-      expect(await utils.getTestAndOverrideConfig(api, triggerConfig, mockReporter, getSummary(), true)).toEqual(
-        expect.objectContaining({
-          test: expect.objectContaining({
-            public_id: '123-456-789',
-            type: 'api',
-            subtype: 'multi',
-            config: {steps: [{subtype: 'http'}, {subtype: 'http'}]},
-          }),
-        })
-      )
-    })
-
-    test('Fails when the tunnel is enabled for an unsupported test type', async () => {
-      const axiosMock = jest.spyOn(axios, 'create')
-      axiosMock.mockImplementation((() => (e: any) => {
-        return {data: {subtype: 'grpc', type: 'api', public_id: '123-456-789'}}
-      }) as any)
-
-      const triggerConfig = {suite: 'Suite 1', config: {}, id: '123-456-789'}
-      await expect(() =>
-        utils.getTestAndOverrideConfig(api, triggerConfig, mockReporter, getSummary(), true)
-      ).rejects.toThrow(
-        'The tunnel is only supported with HTTP API tests and Browser tests (public ID: 123-456-789, type: api, sub-type: grpc).'
-      )
-    })
-
-    test('Fails when the tunnel is enabled for unsupported steps in a Multi step test', async () => {
-      const axiosMock = jest.spyOn(axios, 'create')
-      axiosMock.mockImplementation((() => (e: any) => {
-        return {
-          data: {
-            type: 'api',
-            subtype: 'multi',
-            config: {steps: [{subtype: 'dns'}, {subtype: 'ssl'}, {subtype: 'http'}]},
-            public_id: '123-456-789',
-          },
-        }
-      }) as any)
-
-      const triggerConfig = {suite: 'Suite 1', config: {}, id: '123-456-789'}
-      await expect(() =>
-        utils.getTestAndOverrideConfig(api, triggerConfig, mockReporter, getSummary(), true)
-      ).rejects.toThrow(
-        'The tunnel is only supported with HTTP API tests and Browser tests (public ID: 123-456-789, type: api, sub-type: multi, step sub-types: [dns, ssl]).'
-      )
-    })
-  })
-
   describe('normalizePublicId', () => {
     test('should normalize public ID', () => {
       expect(utils.normalizePublicId('http://localhost/synthetics/tests/details/123-456-789')).toBe('123-456-789')
